requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorlog==6.7.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-extensions==3.2.1
django-filter==23.2
django-phonenumber-field==7.1.0
django-storages==1.13.2
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
graphene==3.2.2
graphene-django==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
mccabe==0.7.0
msgpack==1.0.5
phonenumbers==8.13.14
Pillow==9.5.0
promise==2.3
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pyOpenSSL==23.1.1
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
redis==4.5.5
requests==2.31.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
zope.interface==6.0
